+++ apps/isaaclab.python.kit
@@ -24,8 +24,6 @@ keywords = ["experience", "app", "usd"]
 "isaacsim.gui.menu" = {}
 "isaacsim.gui.property" = {}
 "isaacsim.replicator.behavior" = {}
-"isaacsim.robot_motion.lula" = {}
-"isaacsim.robot_motion.motion_generation" = {}
 "isaacsim.robot.manipulators" = {}
 "isaacsim.robot.policy.examples" = {}
 "isaacsim.robot.schema" = {}
