+++ scripts/imitation_learning/robomimic/play.py
@@ -22,10 +22,6 @@ Args:
 
 import argparse
 
-# Import pinocchio in the main script to force the use of the dependencies installed by IsaacLab and not the one installed by Isaac Sim
-# pinocchio is required by the Pink IK controller
-import pinocchio  # noqa: F401
-
 from isaaclab.app import AppLauncher
 
 # add argparse arguments
@@ -44,12 +40,18 @@ parser.add_argument(
 parser.add_argument(
     "--norm_factor_max", type=float, default=None, help="Optional: maximum value of the normalization factor."
 )
+parser.add_argument("--enable_pinocchio", default=False, action="store_true", help="Enable Pinocchio.")
 
 
 # append AppLauncher cli args
 AppLauncher.add_app_launcher_args(parser)
 # parse the arguments
 args_cli = parser.parse_args()
+
+if args_cli.enable_pinocchio:
+    # Import pinocchio before AppLauncher to force the use of the version installed by IsaacLab and not the one installed by Isaac Sim
+    # pinocchio is required by the Pink IK controllers and the GR1T2 retargeter
+    import pinocchio  # noqa: F401
 
 # launch omniverse app
 app_launcher = AppLauncher(args_cli)
