+++ source/isaaclab/config/extension.toml
@@ -1,7 +1,7 @@
 [package]
 
 # Note: Semantic Versioning is used: https://semver.org/
-version = "0.36.10"
+version = "0.36.11"
 
 # Description
 title = "Isaac Lab framework for Robot Learning"
