+++ source/isaaclab/docs/CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-0.36.10 (2025-04-09)
+0.36.11 (2025-04-09)
 ~~~~~~~~~~~~~~~~~~~~
 
 Changed
@@ -12,8 +12,8 @@ Changed
   the cuda device, which results in NCCL errors on distributed setups.
 
 
-0.36.9 (2025-04-01)
-~~~~~~~~~~~~~~~~~~~
+0.36.10 (2025-04-01)
+~~~~~~~~~~~~~~~~~~~~
 
 Fixed
 ^^^^^
@@ -21,7 +21,7 @@ Fixed
 * Added check in RecorderManager to ensure that the success indicator is only set if the termination manager is present.
 
 
-0.36.8 (2025-03-24)
+0.36.9 (2025-03-24)
 ~~~~~~~~~~~~~~~~~~~
 
 Changed
@@ -31,7 +31,7 @@ Changed
   the default settings will be used from the experience files and the double definition is removed.
 
 
-0.36.7 (2025-03-17)
+0.36.8 (2025-03-17)
 ~~~~~~~~~~~~~~~~~~~
 
 Fixed
@@ -39,6 +39,18 @@ Fixed
 
 * Fixed default behavior of :class:`~isaaclab.actuators.ImplicitActuator` if no :attr:`effort_limits_sim` or
   :attr:`effort_limit` is set.
+
+
+0.36.7 (2025-03-14)
+~~~~~~~~~~~~~~~~~~~
+
+Fixed
+^^^^^
+
+* Changed the import structure to only import ``pinocchio`` when ``pink-ik`` or ``dex-retargeting`` is being used.
+  This also solves for the problem that ``pink-ik`` and ``dex-retargeting`` are not supported in windows.
+* Removed ``isaacsim.robot_motion.lula`` and ``isaacsim.robot_motion.motion_generation`` from the default loaded Isaac Sim extensions.
+* Moved pink ik action config to a separate file.
 
 
 0.36.6 (2025-03-13)
