+++ source/isaaclab/isaaclab/devices/openxr/retargeters/__init__.py
@@ -4,7 +4,6 @@
 # SPDX-License-Identifier: BSD-3-Clause
 """Retargeters for mapping input device data to robot commands."""
 
-from .humanoid.fourier.gr1t2_retargeter import GR1T2Retargeter
 from .manipulator.gripper_retargeter import GripperRetargeter
 from .manipulator.se3_abs_retargeter import Se3AbsRetargeter
 from .manipulator.se3_rel_retargeter import Se3RelRetargeter
