+++ source/isaaclab/isaaclab/envs/mdp/actions/actions_cfg.py
@@ -5,7 +5,7 @@
 
 from dataclasses import MISSING
 
-from isaaclab.controllers import DifferentialIKControllerCfg, OperationalSpaceControllerCfg, PinkIKControllerCfg
+from isaaclab.controllers import DifferentialIKControllerCfg, OperationalSpaceControllerCfg
 from isaaclab.managers.action_manager import ActionTerm, ActionTermCfg
 from isaaclab.utils import configclass
 
@@ -248,30 +248,6 @@ class DifferentialInverseKinematicsActionCfg(ActionTermCfg):
     """Scale factor for the action. Defaults to 1.0."""
     controller: DifferentialIKControllerCfg = MISSING
     """The configuration for the differential IK controller."""
-
-
-@configclass
-class PinkInverseKinematicsActionCfg(ActionTermCfg):
-    """Configuration for Pink inverse kinematics action term.
-
-    This configuration is used to define settings for the Pink inverse kinematics action term,
-    which is a inverse kinematics framework.
-    """
-
-    class_type: type[ActionTerm] = task_space_actions.PinkInverseKinematicsAction
-    """Specifies the action term class type for Pink inverse kinematics action."""
-
-    pink_controlled_joint_names: list[str] = MISSING
-    """List of joint names or regular expression patterns that specify the joints controlled by pink IK."""
-
-    ik_urdf_fixed_joint_names: list[str] = MISSING
-    """List of joint names that specify the joints to be locked in URDF."""
-
-    hand_joint_names: list[str] = MISSING
-    """List of joint names or regular expression patterns that specify the joints controlled by hand retargeting."""
-
-    controller: PinkIKControllerCfg = MISSING
-    """Configuration for the Pink IK controller that will be used to solve the inverse kinematics."""
 
 
 @configclass
