+++ source/isaaclab/isaaclab/envs/mdp/actions/pink_actions_cfg.py
@@ -1,0 +1,36 @@
+# Copyright (c) 2025, The Isaac Lab Project Developers.
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from dataclasses import MISSING
+
+from isaaclab.controllers.pink_ik_cfg import PinkIKControllerCfg
+from isaaclab.managers.action_manager import ActionTerm, ActionTermCfg
+from isaaclab.utils import configclass
+
+from . import pink_task_space_actions
+
+
+@configclass
+class PinkInverseKinematicsActionCfg(ActionTermCfg):
+    """Configuration for Pink inverse kinematics action term.
+
+    This configuration is used to define settings for the Pink inverse kinematics action term,
+    which is a inverse kinematics framework.
+    """
+
+    class_type: type[ActionTerm] = pink_task_space_actions.PinkInverseKinematicsAction
+    """Specifies the action term class type for Pink inverse kinematics action."""
+
+    pink_controlled_joint_names: list[str] = MISSING
+    """List of joint names or regular expression patterns that specify the joints controlled by pink IK."""
+
+    ik_urdf_fixed_joint_names: list[str] = MISSING
+    """List of joint names that specify the joints to be locked in URDF."""
+
+    hand_joint_names: list[str] = MISSING
+    """List of joint names or regular expression patterns that specify the joints controlled by hand retargeting."""
+
+    controller: PinkIKControllerCfg = MISSING
+    """Configuration for the Pink IK controller that will be used to solve the inverse kinematics."""
