+++ source/isaaclab/setup.py
@@ -6,6 +6,7 @@
 """Installation script for the 'isaaclab' python package."""
 
 import os
+import platform
 import toml
 
 from setuptools import setup
@@ -39,9 +40,14 @@ INSTALL_REQUIRES = [
     # livestream
     "starlette==0.46.0",
     "flatdict==4.0.1",
-    "pin-pink==3.1.0",  # required by isaaclab.isaaclab.controllers.pink_ik
-    "dex-retargeting==0.4.6",  # required by isaaclab.devices.openxr.retargeters.humanoid.fourier.gr1_t2_dex_retargeting_utils
 ]
+
+# Additional dependencies that are only available on Linux platforms
+if platform.system() == "Linux":
+    INSTALL_REQUIRES += [
+        "pin-pink==3.1.0",  # required by isaaclab.isaaclab.controllers.pink_ik
+        "dex-retargeting==0.4.6",  # required by isaaclab.devices.openxr.retargeters.humanoid.fourier.gr1_t2_dex_retargeting_utils
+    ]
 
 PYTORCH_INDEX_URL = ["https://download.pytorch.org/whl/cu118"]
