+++ source/isaaclab_tasks/config/extension.toml
@@ -1,7 +1,7 @@
 [package]
 
 # Note: Semantic Versioning is used: https://semver.org/
-version = "0.10.28"
+version = "0.10.29"
 
 # Description
 title = "Isaac Lab Environments"
